schroot-create: drop the whole reproducible setup, actually unneeded
we don't need the reproducible repository in the reproducible schroots,
in the past it could have been (ab)used to test reproduciblity of
packages without actualy uploading, but that's not possible anymore.  It
could have been used to use a diffoscope without uploading to debian,
but that also has only be used once in 2015.

Signed-off-by: Mattia Rizzolo <mattia@debian.org>