schroot-create: revert back to directory-based schroot for reproducible stuff
file-based schroots don't do union mounts, and turns out we would end up
with too many unpacked chroots filling our disk.

This is actually a issue only for diffoscope, but we plan on moving away
from schroot for the other jobs soon.

To still mitigate the concurrency issues on upgrade that we used to
have, save the chroot in a different location each time, and just clean
up old ones in the maintenance jobs.

Signed-off-by: Mattia Rizzolo <mattia@debian.org>