use cdn-fastly directly when proxying through squid
When you configure deb.d.o with apt, apt performs a SRV lookup and
connects to the right backend. When you use a proxy, the proxy is in
charge for performing those SRV lookups. squid doesn't do that. So it
connects to deb.d.o and receives a 302 redirect to cdn-fastly. Except
when it doesn't and receives a timeout instead. This happened a lot
recently.

The whole business of redirecting each request is neither efficient nor
fault tolerant. So we change the default mirror to the default SRV
target when connecting through a squid. Putting the redirects out of the
loop is both faster and avoids the timeouts.