reproducible: Support arbitrary package filters when generating deb822 output.
We were previously encountering errors when filtering excessively-long package
lists. In particular:

  /tmp/jenkins-script-DMnm8bZM: line 25: /usr/bin/grep-dctrl: Argument list too long

This was because we were constructing an excessively-long command-line along
the lines of "(foo|bar|baz...)". This commit changes the implementation to
iterate over Packages file instead to support arbitrary filter sizes.

This is related (but orthogonal...) to MR !15.

Signed-off-by: Holger Levsen <holger@layer-acht.org>