reproducible: health_check: fix a (nonfatal?) syntax error
This was causing:
/tmp/jenkins-script-vkjPKnSH: line 116: "2019" + 1: syntax error: operand expected (error token is ""2019" + 1")

Turns out that within arithmetic variables using $ is discuraged due to
how the expansion is done, and this was indeed leading to a subtle bug
here.
Should have followed shellcheck in the first instance.
https://github.com/koalaman/shellcheck/wiki/SC2004
Curiously, not double-quoting the variables within whould have _also_
prevented this bug (in a different way).

Signed-off-by: Mattia Rizzolo <mattia@debian.org>