packages-tests: include the hashed gitlab token for the lintian-tests_sid job
Otherwise it would be overwritten each time a deployment happen.

Even if the token were to be extracted from the hash and leaked the worst it
would cause is triggering the lintian-tests_sid job, which is not a real
problem for us (only one test at a time can ever run, etc).

Closes: #917119
Signed-off-by: Mattia Rizzolo <mattia@debian.org>