create per distro JSON reproducible info file
The current /debian/reproducible.json file contains other distribution
information as well such as from Arch Linux since the query does not
filter by distro. Resolve this issue by adjusting the SQL query to
filter on the passed distro which defaults to 'debian' if --distro is
not provided as argument. Due to the scripts logic, the created json
files will end up in the distro specific directories.

Signed-off-by: Holger Levsen <holger@layer-acht.org>