Revert "reproducible: add logparse rules to ignore two warnings from unrelated packages installations"
if we have to _ignroe the warnings_ they would need to be "ok" lines,
not "warning", but I can't find a reason as to why they would be
triggered from this logparse rules files, so perhaps it was meant to be
put in another one.

Let me revert this for now.

This reverts commit 99e3da67c9b50eb03a909901f82b21cd8f4423d4.